Release 3.0.2
New option: OPTION series failsafe. When this option is set, Gekko will abort with an error message, when a series statement like for instance y = x; or y[2020] = %v; tries to insert an observation containing a missing value into the left-hand series. The option can be practical for debugging Gekko command files (.gcm), if the source of a missing value is hard to track. The option is only intended for debugging purposes.

Bugfix: The combination of PRT<r> or PRT<m> together with a sum(#i, x[#i]) did not work. The bug is fixed.