New option to control how the main window works as an editor:
  OPTION interface edit style = gekko;  // gekko | gekko2 | rs | rs2

The normal style is 'gekko'. With 'gekko2', a newline will execute the line AND move the cursor to the next line. Therefore, with 'gekko2', multiple lines can be executed with [Enter] [Enter] ... [Enter] instead of [Enter] [downarrow] [Enter] [downarrow] ... [Enter]. In 'gekko' and 'gekko2', [CTRL+Enter] issues a new line without executing anything (like Notepad).

The 'rs' style emulates RStudio, where [Enter] works like Notepad, and [Ctrl+Enter] issues the line. In 'rs', an issued command moves the cursor to the next line (similar to 'gekko2'). With 'rs2', [Ctrl+Enter] does not move to the next line (similar to 'gekko').