Functions arrayPack() and arrayUnpack() can be used to transfer a databank of normal timeseries inside a 1-dimensional array-timeseries, and vice versa.
For IF-ELSE-END structures, there is now also an ELSEIF. This provides nicer syntax in some cases.

"OPTION gams time detect auto = yes" will analyze the parameters/variables of a gdx databank to see if a domain-less dimension (with domain '*') could be a time dimension. This is done by means of "tasting" the data to identify time-like dimensions.